Add Order

+++ Ecommerce/Ecommerce/Models/Product.cs
@@ -9,15 +9,13 @@ namespace Ecommerce.Models
     [Index(nameof(ProductName), IsUnique = true)]
     public class Product : IEntityBase
     {
-        //[Key]
         public int Id { get; set; }
         public string ProductName { get; set; }
         public string Description { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public double PercentSale { get; set; }
         public string ProductImg { get; set; }
         public int Quantity { get; set; }
-
         public int CategoryId { get; set; }
         //[ForeignKey("CategoryId")]
         public virtual Category Category { get; set; }
